+++ src/extension.ts
@@ -1,26 +1,29 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
+import { statusBarItem, modes, currentMode, enterCommands } from "./modes/modes";
 
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
-export function activate(context: vscode.ExtensionContext) {
-  // Use the console to output diagnostic information (console.log) and errors (console.error)
-  // This line of code will only be executed once when your extension is activated
-  console.log('Congratulations, your extension "helloworld-sample" is now active!');
+// TODO check if other extensions are capturing type events
+// TODO add possibility to not activate the extensione if no keybindings are found
+export function activate( context: vscode.ExtensionContext ) {
+    statusBarItem.show();
+    context.subscriptions.push(
+        statusBarItem,
+        // TODO put every command in an array to be disposed afterwards
+        vscode.commands.registerCommand( 'type', keypress => handleKeypress( keypress.text ) ),
+    );
 
-  // The command has been defined in the package.json file
-  // Now provide the implementation of the command with registerCommand
-  // The commandId parameter must match the command field in package.json
-  let disposable = vscode.commands.registerCommand('helloworld.helloWorld', () => {
-    // The code you place here will be executed every time your command is executed
-
-    // Display a message box to the user
-    vscode.window.showInformationMessage('Hello World!');
-  });
-
-  context.subscriptions.push(disposable);
+    vscode.commands.executeCommand( "setContext", "vimcode.active", true );
 }
 
-// this method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() {
+    for( const command of enterCommands ) {
+        command.dispose();
+    }
+
+    statusBarItem.hide();
+    statusBarItem.dispose();
+}
+
+
+function handleKeypress( key: string ): void {
+    modes[ currentMode ].execute( key );
+}