+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 
+## Known Issues
+
+- no settings validation
+
 ## \[Unreleased]
 
-- Default modes: NORMAL, INSERT
-- Ability to define custom modes through the settings.json/settings UI
+- settings validation/error handling
+- type subscription checking
+- adding context settings for each mode, used in keybindings "when" clauses
+
+### 0.0.1
+
+- ability to define custom modes (name, status bar item icon, execution of vscode commands)
+    - unrestricted input: modes that just type the pressed key
+    - restricted input: modes that execute commands